Unbreak CI: There is no such thing as README.md
Rename the script into

```
bin/ensure-practice-exercise-descriptions-are-synced.sh
```

since what was once exercises are now called practice exercises. That
which is called concept exercises is not shared between tracks, and so
syncing their descriptions is not well-defined.

This commit preserves the `override_message`:

```
override_message="I have confirmed ... no README check ..."
```

(censored to avoid triggering CI) in order to not pick a new slogan.

While exercism/configlet#179 is being resolved, the only thing we can
test is that each practice exercise has an instructions.md, which is the
only description file that is always present after #950.