Add single config.json tag to pass 'configlet lint'
To avoid scope creep, adding the full list of tags is addressed in #939.

Note that the formatting of config.json used to be maintained by

```
configlet fmt
```

which is not available in configlet-4.0.0-alpha.4 for v3 tracks.

I've tried to emulate the formatting present in config.json already, but
we currently do not validate formatting in the same way as we used to.